Add Clang AST isomorph → MLIR region mapping infrastructure
Establish methodology for capturing Clang AST graph isomorphisms that
correspond to normalized MLIR base-dialect region expressions, using
hsutter/cppfront as reference transpiler for semantic loss scoring.

Infrastructure:
- corpus/ast_mappings/: N-way mapping artifacts (Cpp2 → C++1 → AST → MLIR)
- tools/generate_ast_mappings.py: Automated corpus processing pipeline
- docs/AST_ISOMORPH_MAPPING_SPEC.md: Complete specification

Example mapping (mixed-hello.cpp2):
- Cpp2 syntax → C++1 translation → Clang AST dump → MLIR Cpp2Dialect
- Documented transformations: function signatures, variable decls, inout
  parameters, UFCS resolution, operator overloading

Semantic loss framework:
- Reference: cppfront transpiler output (ground truth)
- Candidate: cppfort transpiler output (to be validated)
- Metrics: structural distance, type distance, operation distance
- Goal: <0.15 average loss across 195-file corpus

AST isomorph extraction:
- Subgraph pattern identification (functions, control flow, expressions)
- Graph signature computation for isomorphism checking
- Normalization (α-renaming, type canonicalization, CFG reduction)
- MLIR region tagging (cpp2.func, cpp2.if, cpp2.loop, etc.)

MLIR dialect fixes:
- Added BytecodeOpInterface headers for LLVM 21.1.7 compatibility
- Fixed dyn_cast API migration (member → free function)
- Moved mlir_cpp2_dialect.hpp to include/ for proper visibility
- Fixed Scheduler method signatures, Node/Patch default constructors
- Resolved SeaOfNodesBuilder redefinition (header vs implementation)

Build fixes:
- main.cpp: std::println → std::cout/cerr (C++23 → C++20)
- test_scheduler.cpp: Explicit int64_t casts for create_constant()
- sea_of_nodes_ir.cpp: Type::Kind::Struct → UserDefined, TokenType::Star → Asterisk
- tests/CMakeLists.txt: cpp2_transpiler → Cpp2Transpiler library name
- Cpp2Dialect now compiles cleanly

Test status:
- Cpp2Dialect: ✓ builds successfully
- cppfront_regression_tests: ✓ builds, runs (shows parser errors - expected)
- test_scheduler: ⏳ pending (MLIR linker issues - non-critical)

Next steps (tracked in todos):
1. Build hsutter/cppfront from third_party/cppfront/
2. Generate reference AST dumps for 195-file corpus
3. Extract ~1000+ AST isomorphs, tag with MLIR regions
4. Implement semantic loss scoring: cppfort vs cppfront
5. Achieve <0.15 average loss, >40% zero-loss files

Files: 36 changed
- New: 8 (corpus/ast_mappings/*, docs/AST_*.md, tools/generate_ast_mappings.py)
- Modified: 28 (MLIR dialect, transpiler, tests)

Former-commit-id: 584b534d20ed9b82a2d7c126bf976257ca85d303

+++ tests/test_scheduler.cpp
@@ -2,8 +2,8 @@
 #include <iostream>
 
 #include "mlir_cpp2_dialect.hpp"
-#include "sea_of_nodes_ir.cpp"
-#include "pijul_crdt.cpp"
+#include "../src/sea_of_nodes_ir.cpp"
+#include "../src/pijul_crdt.cpp"
 
 int main() {
     using namespace cppfort::mlir_son;
@@ -20,8 +20,8 @@ int main() {
     builder.add_edge(A, B);
 
     // Add data node D that depends on A and B
-    NodeID valA = builder.create_constant(10);
-    NodeID valB = builder.create_constant(20);
+    NodeID valA = builder.create_constant(int64_t(10));
+    NodeID valB = builder.create_constant(int64_t(20));
     NodeID D = builder.create_binary_op(Node::Kind::Add, valA, valB);
     // Make D depend on control B (simulate that both inputs are dominated by B)
     builder.add_edge(B, D);
